+++ src/main/java/HearDay/spring/domain/userarticlebookmark/entity/UserArticleBookmark.java
@@ -1,0 +1,27 @@
+package HearDay.spring.domain.userarticlebookmark.entity;
+
+import HearDay.spring.common.entity.BaseEntity;
+import HearDay.spring.domain.article.entity.Article;
+import HearDay.spring.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class UserArticleBookmark extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "article_id")
+    private Article article;
+}
