+++ src/main/java/HearDay/spring/domain/userarticlebookmark/repository/UserArticleBookmarkRepository.java
@@ -1,0 +1,7 @@
+package HearDay.spring.domain.userarticlebookmark.repository;
+
+import HearDay.spring.domain.userarticlebookmark.entity.UserArticleBookmark;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserArticleBookmarkRepository
+        extends JpaRepository<UserArticleBookmark, Long>, UserArticleBookmarkRepositoryCustom {}
