--- src/main/java/HearDay/spring/domain/wordBookmark/dto/response/WordByDateResponseDto.java
@@ -1,8 +1,0 @@
-package HearDay.spring.domain.wordBookmark.dto.response;
-
-import java.util.List;
-
-public record WordByDateResponseDto(
-    List<WordInfoDto> words
-) {
-}
