--- src/main/java/HearDay/spring/domain/wordBookmark/dto/response/WordCountInfoResponseDto.java
@@ -1,9 +1,0 @@
-package HearDay.spring.domain.wordBookmark.dto.response;
-
-import java.time.LocalDate;
-
-public record WordCountInfoResponseDto(
-        LocalDate date,
-        long count
-) {
-}
