--- src/main/java/HearDay/spring/domain/wordBookmark/dto/response/WordCountResponseDto.java
@@ -1,8 +1,0 @@
-package HearDay.spring.domain.wordBookmark.dto.response;
-
-import java.util.List;
-
-public record WordCountResponseDto(
-        List<WordCountInfoResponseDto> wordCountList
-) {
-}
