--- src/main/java/HearDay/spring/domain/wordBookmark/dto/response/WordDescriptionResponseDto.java
@@ -1,7 +1,0 @@
-package HearDay.spring.domain.wordBookmark.dto.response;
-
-public record WordDescriptionResponseDto(
-        String word,
-        String description
-) {
-}
