--- src/main/java/HearDay/spring/domain/wordBookmark/dto/response/WordInfoDto.java
@@ -1,7 +1,0 @@
-package HearDay.spring.domain.wordBookmark.dto.response;
-
-public record WordInfoDto(
-        Long wordId,
-        String word
-) {
-}
