--- src/main/java/HearDay/spring/domain/wordBookmark/service/wordService/WordCommandService.java
@@ -1,8 +1,0 @@
-package HearDay.spring.domain.wordBookmark.service.wordService;
-
-import HearDay.spring.domain.user.entity.User;
-import HearDay.spring.domain.wordBookmark.dto.request.WordRequestDto;
-
-public interface WordCommandService {
-    void createWordBookmark(User user, WordRequestDto wordRequestDto);
-}
