+++ src/main/java/HearDay/spring/domain/wordbookmark/controller/WordController.java
@@ -1,0 +1,76 @@
+package HearDay.spring.domain.wordbookmark.controller;
+
+import HearDay.spring.common.dto.response.CommonApiResponse;
+import HearDay.spring.domain.user.entity.User;
+import HearDay.spring.domain.wordbookmark.dto.request.WordRequestDto;
+import HearDay.spring.domain.wordbookmark.dto.response.WordByDateResponseDto;
+import HearDay.spring.domain.wordbookmark.dto.response.WordCountResponseDto;
+import HearDay.spring.domain.wordbookmark.dto.response.WordDescriptionResponseDto;
+import HearDay.spring.domain.wordbookmark.service.wordService.WordCommandService;
+import HearDay.spring.domain.wordbookmark.service.wordService.WordQueryService;
+import HearDay.spring.global.annotation.AuthUser;
+import io.swagger.v3.oas.annotations.Operation;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.time.LocalDate;
+
+@RestController
+@RequestMapping("/words")
+@RequiredArgsConstructor
+public class WordController {
+
+    private final WordCommandService wordCommandService;
+    private final WordQueryService wordQueryService;
+
+    @PostMapping("/")
+    @Operation(summary = "단어 저장 API", description = "단어를 내 단어장에 저장시 사용하는 API입니다.")
+    public ResponseEntity<CommonApiResponse<Void>> saveWord(
+            @AuthUser User user,
+            @RequestBody WordRequestDto request
+    ) {
+        wordCommandService.createWordBookmark(user, request);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(CommonApiResponse.success("단어 저장에 성공했습니다.", null));
+    }
+
+    @GetMapping("/{wordsId}")
+    @Operation(summary = "단어 뜻 조회 API", description = "저장된 단어 뜻 확인시 사용하는 API입니다.")
+    public ResponseEntity<CommonApiResponse<WordDescriptionResponseDto>> getWord(
+            @AuthUser User user,
+            @PathVariable Long wordsId
+    ) {
+        WordDescriptionResponseDto result = wordQueryService.getWordDescription(user, wordsId);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(CommonApiResponse.success("단어 뜻 조회에 성공했습니다.", result));
+    }
+
+    @GetMapping("/date")
+    @Operation(summary = "특정 날짜의 단어 목록 조회 API", description = "원하는 날짜를 쿼리 파라미터로 보내면 그날 저장된 단어를 출력하는 API입니다.")
+    public ResponseEntity<CommonApiResponse<WordByDateResponseDto>> getWordDates(
+            @AuthUser User user,
+            @RequestParam LocalDate date
+    ) {
+        WordByDateResponseDto result = wordQueryService.getWordDates(user, date);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(CommonApiResponse.success("해당 날짜의 단어 조회에 성공했습니다.", result));
+    }
+
+    @GetMapping("/statistics")
+    @Operation(summary = "월별 저장된 단어 개수 조회 API", description = "입력한 월별 저장된 단어 개수를 출력하는 API입니다.")
+    public ResponseEntity<CommonApiResponse<WordCountResponseDto>> getWordCounts(
+            @AuthUser User user,
+            @RequestParam int year,
+            @RequestParam int month
+    ) {
+        WordCountResponseDto result = wordQueryService.getWordCounts(user, year, month);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(CommonApiResponse.success("해당 달의 단어 개수 조회에 성공했습니다.", result));
+    }
+}
