+++ src/main/java/HearDay/spring/domain/wordbookmark/dto/request/WordRequestDto.java
@@ -1,0 +1,12 @@
+package HearDay.spring.domain.wordbookmark.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record WordRequestDto(
+        @NotBlank(message = "단어는 필수입니다.")
+        String word,
+
+        @NotBlank(message = "설명은 필수입니다.")
+        String description
+) {
+}
