+++ src/main/java/HearDay/spring/domain/wordbookmark/dto/response/WordByDateResponseDto.java
@@ -1,0 +1,8 @@
+package HearDay.spring.domain.wordbookmark.dto.response;
+
+import java.util.List;
+
+public record WordByDateResponseDto(
+    List<WordInfoDto> words
+) {
+}
