+++ src/main/java/HearDay/spring/domain/wordbookmark/dto/response/WordCountInfoResponseDto.java
@@ -1,0 +1,9 @@
+package HearDay.spring.domain.wordbookmark.dto.response;
+
+import java.time.LocalDate;
+
+public record WordCountInfoResponseDto(
+        LocalDate date,
+        long count
+) {
+}
