+++ src/main/java/HearDay/spring/domain/wordbookmark/dto/response/WordCountResponseDto.java
@@ -1,0 +1,8 @@
+package HearDay.spring.domain.wordbookmark.dto.response;
+
+import java.util.List;
+
+public record WordCountResponseDto(
+        List<WordCountInfoResponseDto> wordCountList
+) {
+}
