+++ src/main/java/HearDay/spring/domain/wordbookmark/dto/response/WordDescriptionResponseDto.java
@@ -1,0 +1,7 @@
+package HearDay.spring.domain.wordbookmark.dto.response;
+
+public record WordDescriptionResponseDto(
+        String word,
+        String description
+) {
+}
