+++ src/main/java/HearDay/spring/domain/wordbookmark/dto/response/WordInfoDto.java
@@ -1,0 +1,7 @@
+package HearDay.spring.domain.wordbookmark.dto.response;
+
+public record WordInfoDto(
+        Long wordId,
+        String word
+) {
+}
