+++ src/main/java/HearDay/spring/domain/wordbookmark/entity/UserWordBookmark.java
@@ -1,0 +1,28 @@
+package HearDay.spring.domain.wordbookmark.entity;
+
+import HearDay.spring.common.entity.BaseEntity;
+import HearDay.spring.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class UserWordBookmark extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String word;
+
+    @Column(nullable = false)
+    private String description;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+}
